MsbmsSystError.Types (msbms_syst_error v0.1.0)

Defines the public types exposed by the MsbmsSystError component.

Types

msbms_error()

@type msbms_error() ::
  :undefined_error
  | :database_error
  | :example_error
  | :file_error
  | :invalid_data
  | :invalid_parameter
  | :permission_denied
  | :record_not_found
  | :process_error

Categories or groups of error kinds which might be seen in the application.

The level of granularity we're trying to achieve here is we want any logic that makes decisions on whether to handle an error/exception to be able to understand the nature of the error.